+++ internal/provider/cloudgatewaytransitgateway_resource.go
@@ -217,7 +217,7 @@ func (r *CloudGatewayTransitGatewayResource) Schema(ctx context.Context, req res
 						MarkdownDescription: `The current state of the Transit Gateway. Possible values:` + "\n" +
 							`- ` + "`" + `created` + "`" + ` - The attachment has been created but is not attached to transit gateway.` + "\n" +
 							`- ` + "`" + `initializing` + "`" + ` - The attachment is in the process of being initialized and is setting up necessary resources.` + "\n" +
-							`- ` + "`" + `pending` + "`" + ` - acceptance The attachment request is awaiting acceptance in customer VPC.` + "\n" +
+							`- ` + "`" + `pending-acceptance` + "`" + ` The attachment request is awaiting acceptance in customer VPC.` + "\n" +
 							`- ` + "`" + `ready` + "`" + ` - The transit gateway attachment is fully operational and can route traffic as configured.` + "\n" +
 							`- ` + "`" + `terminating` + "`" + ` - The attachment is in the process of being deleted and is no longer accepting new traffic.` + "\n" +
 							`- ` + "`" + `terminated` + "`" + ` - The attachment has been fully deleted and is no longer available.` + "\n" +
@@ -437,7 +437,7 @@ func (r *CloudGatewayTransitGatewayResource) Schema(ctx context.Context, req res
 						MarkdownDescription: `The current state of the Transit Gateway. Possible values:` + "\n" +
 							`- ` + "`" + `created` + "`" + ` - The attachment has been created but is not attached to transit gateway.` + "\n" +
 							`- ` + "`" + `initializing` + "`" + ` - The attachment is in the process of being initialized and is setting up necessary resources.` + "\n" +
-							`- ` + "`" + `pending` + "`" + ` - acceptance The attachment request is awaiting acceptance in customer VPC.` + "\n" +
+							`- ` + "`" + `pending-acceptance` + "`" + ` The attachment request is awaiting acceptance in customer VPC.` + "\n" +
 							`- ` + "`" + `ready` + "`" + ` - The transit gateway attachment is fully operational and can route traffic as configured.` + "\n" +
 							`- ` + "`" + `terminating` + "`" + ` - The attachment is in the process of being deleted and is no longer accepting new traffic.` + "\n" +
 							`- ` + "`" + `terminated` + "`" + ` - The attachment has been fully deleted and is no longer available.` + "\n" +
@@ -649,7 +649,7 @@ func (r *CloudGatewayTransitGatewayResource) Schema(ctx context.Context, req res
 						MarkdownDescription: `The current state of the Transit Gateway. Possible values:` + "\n" +
 							`- ` + "`" + `created` + "`" + ` - The attachment has been created but is not attached to transit gateway.` + "\n" +
 							`- ` + "`" + `initializing` + "`" + ` - The attachment is in the process of being initialized and is setting up necessary resources.` + "\n" +
-							`- ` + "`" + `pending` + "`" + ` - acceptance The attachment request is awaiting acceptance in customer VPC.` + "\n" +
+							`- ` + "`" + `pending-acceptance` + "`" + ` The attachment request is awaiting acceptance in customer VPC.` + "\n" +
 							`- ` + "`" + `ready` + "`" + ` - The transit gateway attachment is fully operational and can route traffic as configured.` + "\n" +
 							`- ` + "`" + `terminating` + "`" + ` - The attachment is in the process of being deleted and is no longer accepting new traffic.` + "\n" +
 							`- ` + "`" + `terminated` + "`" + ` - The attachment has been fully deleted and is no longer available.` + "\n" +